# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */
export interface Ok {
  ok: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  clerkId?: string;
  username: string;
  displayName: string;
  bio: string;
  profilePicture: string;
  isVerified: boolean;
  verificationRequested: boolean;
  followersCount: number;
  followingCount: number;
  videosCount: number;
  lastSeenVisible: boolean;
  createdAt: string;
}

export interface UserSummary {
  id: number;
  username: string;
  displayName: string;
  bio: string;
  profilePicture: string;
  isVerified: boolean;
  followersCount: number;
  isFollowing: boolean;
}

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  bio: string;
  profilePicture: string;
  isVerified: boolean;
  followersCount: number;
  followingCount: number;
  videosCount: number;
  likesCount: number;
  isFollowing: boolean;
  isMe: boolean;
}

export interface FollowState {
  isFollowing: boolean;
  followersCount: number;
}

export interface Video {
  id: number;
  videoUrl: string;
  thumbnailUrl: string;
  caption: string;
  likesCount: number;
  commentsCount: number;
  liked: boolean;
  createdAt: string;
  author: UserSummary;
}

export interface LikeState {
  liked: boolean;
  likesCount: number;
}

export interface Comment {
  id: number;
  body: string;
  createdAt: string;
  author: UserSummary;
}

export type ChatPreviewLastMessageKind =
  (typeof ChatPreviewLastMessageKind)[keyof typeof ChatPreviewLastMessageKind];

export const ChatPreviewLastMessageKind = {
  text: "text",
  voice: "voice",
  emoji: "emoji",
  image: "image",
} as const;

export type ChatPreviewLastMessageStatus =
  (typeof ChatPreviewLastMessageStatus)[keyof typeof ChatPreviewLastMessageStatus];

export const ChatPreviewLastMessageStatus = {
  sent: "sent",
  delivered: "delivered",
  read: "read",
} as const;

export interface ChatPreview {
  partner: UserSummary;
  lastMessage: string;
  lastMessageKind: ChatPreviewLastMessageKind;
  lastMessageAt: string;
  lastMessageFromMe: boolean;
  lastMessageStatus: ChatPreviewLastMessageStatus;
  unreadCount: number;
  partnerTyping: boolean;
  partnerOnline: boolean;
  pinned: boolean;
  archived: boolean;
  blocked: boolean;
}

export type MessageKind = (typeof MessageKind)[keyof typeof MessageKind];

export const MessageKind = {
  text: "text",
  voice: "voice",
  emoji: "emoji",
  image: "image",
} as const;

export type MessageStatus = (typeof MessageStatus)[keyof typeof MessageStatus];

export const MessageStatus = {
  sent: "sent",
  delivered: "delivered",
  read: "read",
} as const;

export interface Message {
  id: number;
  senderUsername: string;
  receiverUsername: string;
  body: string;
  kind: MessageKind;
  status: MessageStatus;
  deletedForEveryone: boolean;
  createdAt: string;
}

export type StatusKind = (typeof StatusKind)[keyof typeof StatusKind];

export const StatusKind = {
  text: "text",
  image: "image",
  video: "video",
} as const;

export interface Status {
  id: number;
  kind: StatusKind;
  body: string;
  mediaUrl: string;
  backgroundColor: string;
  createdAt: string;
  expiresAt: string;
}

export interface StatusGroup {
  author: UserSummary;
  statuses: Status[];
  isMe: boolean;
}

export interface BlockState {
  isBlocked: boolean;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  follow: "follow",
  like: "like",
  comment: "comment",
  message: "message",
  verified: "verified",
} as const;

export interface Notification {
  id: number;
  type: NotificationType;
  message: string;
  isRead: boolean;
  createdAt: string;
  actor?: UserSummary;
  videoId?: number;
}

export interface UnreadCount {
  count: number;
}

export type UpdateMeBody = {
  displayName?: string;
  bio?: string;
  profilePicture?: string;
};

export type SearchUsersParams = {
  q: string;
};

export type UpdatePrivacyBody = {
  lastSeenVisible?: boolean;
};

export type CreateStatusBodyKind =
  (typeof CreateStatusBodyKind)[keyof typeof CreateStatusBodyKind];

export const CreateStatusBodyKind = {
  text: "text",
  image: "image",
  video: "video",
} as const;

export type CreateStatusBody = {
  kind: CreateStatusBodyKind;
  body?: string;
  mediaUrl?: string;
  backgroundColor?: string;
};

export type CreateVideoBody = {
  videoUrl: string;
  thumbnailUrl: string;
  caption: string;
};

export type AddCommentBody = {
  body: string;
};

export type GetChatsParams = {
  /**
   * Pass "true" to fetch archived chats only. Otherwise non-archived chats are returned.
   */
  archived?: GetChatsArchived;
};

export type GetChatsArchived =
  (typeof GetChatsArchived)[keyof typeof GetChatsArchived];

export const GetChatsArchived = {
  true: "true",
  false: "false",
} as const;

export type SendMessageBodyKind =
  (typeof SendMessageBodyKind)[keyof typeof SendMessageBodyKind];

export const SendMessageBodyKind = {
  text: "text",
  voice: "voice",
  emoji: "emoji",
  image: "image",
} as const;

export type SendMessageBody = {
  body: string;
  kind: SendMessageBodyKind;
  clientId?: string;
};

export type DeleteMessageParams = {
  scope?: DeleteMessageScope;
};

export type DeleteMessageScope =
  (typeof DeleteMessageScope)[keyof typeof DeleteMessageScope];

export const DeleteMessageScope = {
  me: "me",
  everyone: "everyone",
} as const;