# lib/api-zod/src/generated/types/status.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StatusKind } from "./statusKind";

export interface Status {
  id: number;
  kind: StatusKind;
  body: string;
  mediaUrl: string;
  backgroundColor: string;
  createdAt: Date;
  expiresAt: Date;
}