# lib/api-zod/src/generated/types/statusGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Status } from "./status";
import type { UserSummary } from "./userSummary";

export interface StatusGroup {
  author: UserSummary;
  statuses: Status[];
  isMe: boolean;
}