# lib/api-zod/src/generated/types/userSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kizchat social platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSummary {
  id: number;
  username: string;
  displayName: string;
  bio: string;
  profilePicture: string;
  isVerified: boolean;
  followersCount: number;
  isFollowing: boolean;
}